Refactored the text-based script parsing into a custom type-checked script interpreter
- Database loading errors now exit the game / editor
- A single IScriptContext interface replaces the global monobehaviours
passed around
- Variables and Functions are shared through the IScriptContext
- Global Variables are Variables determined from the Context (ex :
company money)
- Strongly typed Variables (no type inferring nor conversions for the
moment) : bool, int, float, id, string, date ; fonctions are operations
are storngly typed too
- Expression Parsing : done with the Shunting-Yard algorithm : much
better results than with string manual string splits

+++ Assets/Actors/EmployeeSkill.cs
@@ -10,10 +10,10 @@ public class EmployeeSkill {
     [SerializeField] private string name;
     public string Name => name;
 
-    [SerializeField] private int proficiency;
-    public int Proficiency => proficiency;
+    [SerializeField] private float proficiency;
+    public float Proficiency => proficiency;
 
-    public EmployeeSkill(string id, string name, int proficiency) {
+    public EmployeeSkill(string id, string name, float proficiency) {
         Assert.IsTrue(proficiency >= 0);
         this.id = id;
         this.name = name;
